Fixed colspan="1" being applied to all HTML tables
The "span" attribute of StyleAttributes returns 1 by default so that it
can be used in arithmetic elsewhere. That made HTMLTable emit 'colspan="1"'
for every td. That's not technically incorrect, but it's ugly and
unnecessary.

Finally got requirements working in setup.py. I wish that was better
documented.

Hid some distutils files from git.