Added hooks for custom casting functions
Added the 'castfunctions' class variable that holds a dict of types and
the functions used to cast them to strings (default is str() ). For example,
I use this function on a website to print datetime objects pulled from a
database in the US/Central timezone and in the format customers are used to:

    central = pytz.timezone('US/Central')
    def formatdatetime(dtobj):
        if formatdatetime is None:
            return ''
        try:
            centraldate = dtobj.astimezone(central)
        except ValueError:
            centraldate = dtobj.replace(tzinfo=pytz.utc).astimezone(central)
        return centraldate.strftime('%m/%d/%Y %H:%M:%S')

    [...]

    reportmaker = HTMLTable(title, explanation, headers=rowspec)
    reportmaker.castfunctions[datetime.datetime] = formatdatetime
    report = reportmaker.render(rows)

Updated the text to reflect the correct (MIT) license.